+++ R/betafunctions.R
@@ -1069,7 +1069,7 @@ gchoose <- function(n, k) {
 #' # Assume some variable follows a Gamma-Binomial  distribution with
 #' # "number of trials" = 10.5 and probability of "success" for each "trial"
 #' # = 0.75, to compute the cumulative probability to attain a "number of
-#' success" below a specific point (e.g., less than 7.5 "successes":
+#' # success" below a specific point (e.g., less than 7.5 "successes":
 #' pGammaBinom(q = 7.5, size = 10.5, prob = 0.75)
 #'
 #' # Conversely, to attain a value at or above 7.5: